Make bracket initialization persistent
The problem arose from editing and saving the bracket correctly in localStorage, but not updating the localStorage's version of the tournament. This resulted in an outdated localStorage tournament, which would then overwrite a correctly saved one. BE CAREFUL with localStorage. In fact, it might be best to conglomerate all localStorages into one string so that decoupled localStorage calls like this don't cause problems again. Generally speaking, the storage flow should be more well defined to increase readability and maintainability.